requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
branca==0.4.1
certifi==2020.6.20
cffi==1.14.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-tinymce4-lite==1.8.0
filelock==3.0.12
folium==0.11.0
gunicorn==20.1.0
identify==1.5.0
idna==2.10
isort==4.3.21
Jinja2==2.11.2
jsmin==2.2.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mysqlclient @ file:///C:/Users/HP/Blog%20App%20Django/blogapp/mysqlclient-1.4.6-cp38-cp38-win32.whl
nodeenv==1.5.0
numpy==1.19.0
oauthlib==3.1.1
packaging==20.4
Pillow==8.1.0
pluggy==0.13.1
pre-commit==2.7.1
psycopg2==2.8.6
py==1.9.0
PyAutoGUI==0.9.52
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.9
PyJWT==2.1.0
pylint==2.5.3
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2020.1
pywhatkit==3.8
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
toml==0.10.1
tox==3.20.0
translate==3.5.0
urllib3==1.25.10
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wikipedia==1.4.0
wrapt==1.12.1
